+++ examples/src/main/java/org/apache/ignite/examples/ml/tutorial/Step_6_KNN.java
@@ -21,7 +21,7 @@ import java.io.FileNotFoundException;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.IgniteCache;
 import org.apache.ignite.Ignition;
-import org.apache.ignite.ml.knn.classification.KNNClassificationModel;
+import org.apache.ignite.ml.knn.NNClassificationModel;
 import org.apache.ignite.ml.knn.classification.KNNClassificationTrainer;
 import org.apache.ignite.ml.knn.classification.KNNStrategy;
 import org.apache.ignite.ml.math.functions.IgniteBiFunction;
@@ -88,7 +88,7 @@ public class Step_6_KNN {
                     KNNClassificationTrainer trainer = new KNNClassificationTrainer();
 
                     // Train decision tree model.
-                    KNNClassificationModel mdl = trainer.fit(
+                    NNClassificationModel mdl = trainer.fit(
                         ignite,
                         dataCache,
                         normalizationPreprocessor,
