+++ modules/ml/src/main/java/org/apache/ignite/ml/knn/ann/ANNClassificationModel.java
@@ -15,24 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.ml.knn.classification;
+package org.apache.ignite.ml.knn.ann;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
-import org.apache.ignite.ml.Exportable;
-import org.apache.ignite.ml.Exporter;
-import org.apache.ignite.ml.Model;
-import org.apache.ignite.ml.dataset.Dataset;
-import org.apache.ignite.ml.dataset.primitive.context.EmptyContext;
-import org.apache.ignite.ml.math.distances.DistanceMeasure;
-import org.apache.ignite.ml.math.distances.EuclideanDistance;
+import org.apache.ignite.ml.knn.NNClassificationModel;
+import org.apache.ignite.ml.knn.classification.KNNStrategy;
 import org.apache.ignite.ml.math.primitives.vector.Vector;
 import org.apache.ignite.ml.structures.LabeledDataset;
 import org.apache.ignite.ml.structures.LabeledVector;
@@ -40,87 +33,44 @@ import org.apache.ignite.ml.util.ModelTrace;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * kNN algorithm model to solve multi-class classification task.
+ * ANN model to predict labels in multi-class classification task.
  */
-public class KNNWithACDClassificationModel implements Model<Vector, Double>, Exportable<KNNModelFormat> {
+public class ANNClassificationModel extends NNClassificationModel implements Exportable<ANNModelFormat> {
     /** */
-    private static final long serialVersionUID = -127386523291350345L;
+    private static final long serialVersionUID = -127312378991350345L;
 
     private final LabeledDataset<ProbableLabel, LabeledVector> candidates;
 
-    /** Amount of nearest neighbors. */
-    protected int k = 5;
-
-    /** Distance measure. */
-    protected DistanceMeasure distanceMeasure = new EuclideanDistance();
-
-    /** kNN strategy. */
-    protected KNNStrategy stgy = KNNStrategy.SIMPLE;
-
-
-    public KNNWithACDClassificationModel(LabeledDataset<ProbableLabel, LabeledVector> centers) {
+    /**
+     * Build the model based on a candidates set.
+     * @param centers The candidates set.
+     */
+    public ANNClassificationModel(LabeledDataset<ProbableLabel, LabeledVector> centers) {
        this.candidates = centers;
     }
 
     /** {@inheritDoc} */
     @Override public Double apply(Vector v) {
-
             List<LabeledVector> neighbors = findKNearestNeighbors(v);
-
             return classify(neighbors, v, stgy);
-
     }
 
     /** */
-    @Override public <P> void saveModel(Exporter<KNNModelFormat, P> exporter, P path) {
-        KNNModelFormat mdlData = new KNNModelFormat(k, distanceMeasure, stgy);
+    @Override public <P> void saveModel(Exporter<ANNModelFormat, P> exporter, P path) {
+        ANNModelFormat mdlData = new ANNModelFormat(k, distanceMeasure, stgy);
         exporter.save(mdlData, path);
     }
 
     /**
-     * Set up parameter of the kNN model.
-     * @param k Amount of nearest neighbors.
-     * @return Model.
-     */
-    public KNNWithACDClassificationModel withK(int k) {
-        this.k = k;
-        return this;
-    }
-
-    /**
-     * Set up parameter of the kNN model.
-     * @param stgy Strategy of calculations.
-     * @return Model.
-     */
-    public KNNWithACDClassificationModel withStrategy(KNNStrategy stgy) {
-        this.stgy = stgy;
-        return this;
-    }
-
-    /**
-     * Set up parameter of the kNN model.
-     * @param distanceMeasure Distance measure.
-     * @return Model.
-     */
-    public KNNWithACDClassificationModel withDistanceMeasure(DistanceMeasure distanceMeasure) {
-        this.distanceMeasure = distanceMeasure;
-        return this;
-    }
-
-    /**
-     * The main idea is calculation all distance pairs between given vector and all vectors in training set, sorting
+     * The main idea is calculation all distance pairs between given vector and all centroids in candidates set, sorting
      * them and finding k vectors with min distance with the given vector.
      *
      * @param v The given vector.
      * @return K-nearest neighbors.
      */
-    protected List<LabeledVector> findKNearestNeighbors(Vector v) {
-
-
+    private List<LabeledVector> findKNearestNeighbors(Vector v) {
         return Arrays.asList(getKClosestVectors(getDistances(v)));
     }
-
-
 
     /**
      * Iterates along entries in distance map and fill the resulting k-element array.
@@ -177,19 +127,6 @@ public class KNNWithACDClassificationModel implements Model<Vector, Double>, Exp
     }
 
     /** */
-    private void putDistanceIdxPair(Map<Double, Set<Integer>> distanceIdxPairs, int i, double distance) {
-        if (distanceIdxPairs.containsKey(distance)) {
-            Set<Integer> idxs = distanceIdxPairs.get(distance);
-            idxs.add(i);
-        }
-        else {
-            Set<Integer> idxs = new HashSet<>();
-            idxs.add(i);
-            distanceIdxPairs.put(distance, idxs);
-        }
-    }
-
-    /** */
     private double classify(List<LabeledVector> neighbors, Vector v, KNNStrategy stgy) {
         Map<Double, Double> clsVotes = new HashMap<>();
 
@@ -198,45 +135,13 @@ public class KNNWithACDClassificationModel implements Model<Vector, Double>, Exp
 
             double distance = distanceMeasure.compute(v, neighbor.features());
 
-            // let's think
+            // we predict class label, not the probability vector (it need here another math with counting of votes)
             probableClsLb.forEach((label, probability) -> {
-
                 double cnt = clsVotes.containsKey(label) ? clsVotes.get(label) : 0;
-                centroidStat.put(lb, cnt + 1);
-
-
+                clsVotes.put(label, cnt + probability * getClassVoteForVector(stgy, distance));
             });
-
-            for (int clsIdx = 0; clsIdx < probableClsLb.length; clsIdx++) {
-
-                final double clsLbl = clsIdx + 1;
-                if (clsVotes.containsKey(clsLbl)) {
-                    double clsVote = clsVotes.get(clsLbl);
-                    clsVote += getClassVoteForVector(stgy, distance) * probableClsLb[clsIdx];
-                    clsVotes.put(clsLbl, clsVote);
-                }
-                else {
-                    final double val = getClassVoteForVector(stgy, distance);
-                    clsVotes.put(clsLbl, val);
-                }
-
-            }
-
         }
         return getClassWithMaxVotes(clsVotes);
-    }
-
-    /** */
-    private double getClassWithMaxVotes(Map<Double, Double> clsVotes) {
-        return Collections.max(clsVotes.entrySet(), Map.Entry.comparingByValue()).getKey();
-    }
-
-    /** */
-    private double getClassVoteForVector(KNNStrategy stgy, double distance) {
-        if (stgy.equals(KNNStrategy.WEIGHTED))
-            return 1 / distance; // strategy.WEIGHTED
-        else
-            return 1.0; // strategy.SIMPLE
     }
 
     /** {@inheritDoc} */
@@ -246,6 +151,7 @@ public class KNNWithACDClassificationModel implements Model<Vector, Double>, Exp
         res = res * 37 + k;
         res = res * 37 + distanceMeasure.hashCode();
         res = res * 37 + stgy.hashCode();
+        res = res * 37 + candidates.hashCode();
 
         return res;
     }
@@ -258,9 +164,12 @@ public class KNNWithACDClassificationModel implements Model<Vector, Double>, Exp
         if (obj == null || getClass() != obj.getClass())
             return false;
 
-        KNNWithACDClassificationModel that = (KNNWithACDClassificationModel)obj;
+        ANNClassificationModel that = (ANNClassificationModel)obj;
 
-        return k == that.k && distanceMeasure.equals(that.distanceMeasure) && stgy.equals(that.stgy);
+        return k == that.k
+            && distanceMeasure.equals(that.distanceMeasure)
+            && stgy.equals(that.stgy)
+            && candidates.equals(that.candidates);
     }
 
     /** {@inheritDoc} */
@@ -274,6 +183,7 @@ public class KNNWithACDClassificationModel implements Model<Vector, Double>, Exp
             .addField("k", String.valueOf(k))
             .addField("measure", distanceMeasure.getClass().getSimpleName())
             .addField("strategy", stgy.name())
+            .addField("amount of candidates", String.valueOf(candidates.rowSize()))
             .toString();
     }
 }
