--- modules/ml/src/main/java/org/apache/ignite/ml/knn/classification/KNNGraph.java
@@ -1,8 +1,0 @@
-package org.apache.ignite.ml.knn.classification;
-
-
-import java.util.Map;
-
-public class KNNGraph {
-    private Map map;
-}
